+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+* Add required parameter `program_name` to `Fui::new()` - Breaking change
+* Ensure `Fui` action names are unique or panic
 * `Form` can be dumped to `CLI` command by `ctrl+k`
 * Cancel `Fui`'s form shows back action picker
 * Update to new `Cursive` `API` - Breaking change